Hatfield/bala/studentmodels (#4)
* Tentatively add fields to Student model

For a couple of fields that need more info to implement, max_length=32
was used. Also, placeholders for the unimplemented Class and Grade
models were used. This has not been tested.

* Add blanks & defaults to model fields

This was to fix the problem where django did not know how to populate
existing entries when making migrations. We decided that we can validate
these somewhere else anyways, if we don't want some fields to be
blank/default.

* Fix import errors

* Create serializer for student